// src/front/src/api/rcp/models/rcpTimeEntryStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Rcp API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export type RcpTimeEntryStatus =
	(typeof RcpTimeEntryStatus)[keyof typeof RcpTimeEntryStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RcpTimeEntryStatus = {
	Draft: "Draft",
	Submitted: "Submitted",
	Approved: "Approved",
	Correction: "Correction",
	Settlement: "Settlement",
} as const;